Crate generic_mutability_test

Example of restricting statically-checked mutability for types, that are not originally designed for this (e.g. FFI wrappers etc) or which uses interior mutability, but still need verbose API interface.

The problem

Let’s start from the monkey_ffi module.

This is an example of extern C API, that looks like object-oriented, but is not safe and not guarantees the existence of objects, as well as their relations.

It looks like tree-structure:

Root
----Window
----Frame
----|----FrameButton
----WindowButton

Ideally, only one object at time should be mutable, and we have to guarantee, that parent lives at least as long as child.

But the problem is, that with Rc, or just pointers, on every fold level parent mutability state is lost. We can not make easily two versions of Window: one, that keeps &'a mut Root and other, that keeps &'a Root, just because we will need to type whole the implementation block twice.

But the fact, that Self, &Self and &mut Self are complete different types and this discussion pointed me the solution.

Actually, mutability in rust is tri-state: immutable, mutable, and that, which we don’t care about. Which we want to be «generic». So, let’s start from declaring types, that will represent all the three states: two concrete structs and one trait.

trait ProbablyMutable {
fn is_mutable(&self) -> bool;
}
struct Mutable;
impl ProbablyMutable for Mutable {
    fn is_mutable(&self) -> bool {
        true
    }
}
struct Immutable;
impl ProbablyMutable for Immutable {
    fn is_mutable(&self) -> bool {
        false
    }
}

Then we will use these types as markers for future parametrization.

Now let’s make a skeleton of object structure and ensure that no child will outlive their parents.

struct Root;

Here we use PhantomData to keep «generic» part outside of any concrete parent object. Later, it will help not to have in a deep children structure scary constructions, like: SecondChild<&mut Parent, &mut FirstChild<&mut Parent>>

struct Window<'a, T: ProbablyMutable> {
    id: usize,
    name: String,
    frames_amount: usize,
    buttons_amount: usize,
    root: &'a Root,
    mutability: PhantomData<T>,
}

struct Frame<'a, T: ProbablyMutable> {
    window: &'a Window<'a, T>,
    id: usize,
    width_px: Option<u16>,
    buttons_amount: usize,
}

struct WindowButton<'a, T: ProbablyMutable> {
    id: usize,
    text: String,
    parent: &'a Window<'a, T>,
}

struct FrameButton<'a, T: ProbablyMutable> {
    id: usize,
    text: String,
    parent: &'a Frame<'a, T>,
}

since there are two different functions sets for buttons I decided to consider them as different classes with the single interface (trait) Button.

Theoretically, it should be possible to make a single struct, that keeps differs of types in enum. But for the moment, it seemed to me like an overhead.

Considering the generic mutation, I implement it like implementation of three different types:

struct<T: ProbablyMutable> for functions, that should be generic in their mutability.

struct<Mutable> for functions, that require object to be mutable.

struct<Immutable> for functions, that require object to be immutable.

impl<'a, T: ProbablyMutable> Window<'a, T> {
    fn new(root: &'a Root, id: usize) -> Option<Self> {todo!()}
    fn get_id(&self) -> usize {todo!()}
    fn get_name(&self) -> &String {todo!()}
    fn get_width(&self) -> u16 {todo!()}
}
impl<'a> Window<'a, Immutable> {
    fn get_frame(&self, id: usize) -> Option<Frame<Immutable>> {todo!()}
    fn get_button(&self, id: usize) -> Option<WindowButton<Immutable>> {todo!()}
}
impl<'a> Window<'a, Mutable> {
    fn set_name(&mut self, name: impl Into<String>) {todo!()}
    fn make_frame(&mut self) -> Frame<Mutable> {todo!()}
    fn make_button(&mut self) -> WindowButton<Mutable> {todo!()}
}

Modules

monkey_ffi 🔒

module that imitates some FFI functions set.

Structs

Frame 🔒

FrameButton 🔒

Immutable 🔒

Mutable 🔒

Root 🔒

Window 🔒

WindowButton 🔒

Traits

Button 🔒

ButtonMut 🔒

ProbablyMutable 🔒

Functions

main 🔒